requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
numpy
PyPDF2
python-docx
lxml
requests
python-multipart
markdown
beautifulsoup4
openpyxl
sqlalchemy
pillow
python-dateutil
pytz
tzdata